Make demographics survey configurable
The demographics survey of the questionnaire is the most likely to
change over time. First, the questions are not the exact same as the
ones in the original paper. Instead, these have been converted to
questions that are more familiar to US citizens since the original ones
were geared more towards British (I believe). In addition to refining
the current set of question, I could envision adding more over time as
I come up with more facets that I would like to slice the data on. For
this reason, I decided to make the demographics questions configurable.
I've added a version number, just in case questions are modified (e.g.,
the selection 4 for ethnicity could from version to version). Also, it
might be worth considering localizing the questions as well and making
this configurable would help with that considerably.